camera: fix ownership of CameraText
A CameraText is basically a typedef for an array of chars. We can allocate that
and have the library fill our structure, and we can get a string out of it.

However, the existing code takes the buffer as a `Vec<u8>` and converts it into
a `String`. This will cause us to try to free the memory pointed to by the
vector when the string goes out of scope. This means we'll try freeing something
that was allocated on the stack.

Instead, allocate on the heap and let the `String` properly take ownership of
the memory.